requirements.txt
opencv-python
numpy
Pillow
pyautogui
pygetwindow
keyboard
customtkinter
easyocr
mss
psutil
pywin32
setuptools
Cython
pyinstaller